Fix bytes calculation in x16-packw benchmark
- sizeof(uint16_t) replaces sizeof(float)

PiperOrigin-RevId: 696036622

+++ bench/packw-benchmark.h
@@ -188,7 +188,7 @@ static void x16_packw(benchmark::State& state,
   state.counters["elements"] =
     benchmark::Counter(uint64_t(state.iterations()) * elements_per_iteration, benchmark::Counter::kIsRate);
 
-  const size_t bytes_per_iteration = (elements_per_iteration + batch * (rounded_n * rounded_k + rounded_n)) * sizeof(float);
+  const size_t bytes_per_iteration = (elements_per_iteration + batch * (rounded_n * rounded_k + rounded_n)) * sizeof(uint16_t);
   state.counters["bytes"] =
     benchmark::Counter(uint64_t(state.iterations()) * bytes_per_iteration, benchmark::Counter::kIsRate);
 }
